Option stuff
Renamed input options to tsx files
