--- devfiles/src/classes/options/InputOption.ts
@@ -1,5 +1,0 @@
-import Query from '../query/Query';
-
-export default abstract class InputOption {
-    abstract query(): Query;
-}
